refactor: avoid cycle dependencies
