+++ src/tx/builder/helpers.js
@@ -11,7 +11,7 @@ import {
   NAME_BID_TIMEOUTS,
   NAME_MAX_LENGTH_FEE,
   POINTER_KEY_BY_PREFIX
-} from './schema'
+} from './constants'
 import { ceil } from '../../utils/bignumber'
 import {
   TagNotFoundError,
