+++ src/tx/builder/schema.js
@@ -10,6 +10,7 @@
 import BigNumber from 'bignumber.js'
 import { Name, NameId, NameFee, Deposit } from './field-types'
 
+export * from './constants'
 export const VSN = 1
 export const VSN_2 = 2
 
@@ -26,83 +27,6 @@ export const GAS_MAX = 1600000 - 21000
 export const MIN_GAS_PRICE = 1e9
 export const MAX_AUTH_FUN_GAS = 50000
 export const DRY_RUN_ACCOUNT = { pub: 'ak_11111111111111111111111111111111273Yts', amount: '100000000000000000000000000000000000' }
-// # AENS
-export const NAME_TTL = 180000
-// # max number of block into the future that the name is going to be available
-// # https://github.com/aeternity/protocol/blob/epoch-v0.22.0/AENS.md#update
-// # https://github.com/aeternity/protocol/blob/44a93d3aab957ca820183c3520b9daf6b0fedff4/AENS.md#aens-entry
-export const NAME_MAX_TTL = 36000
-export const NAME_MAX_CLIENT_TTL = 84600
-export const CLIENT_TTL = NAME_MAX_CLIENT_TTL
-// # see https://github.com/aeternity/aeternity/blob/72e440b8731422e335f879a31ecbbee7ac23a1cf/apps/aecore/src/aec_governance.erl#L67
-export const NAME_FEE_MULTIPLIER = 1e14 // 100000000000000
-export const NAME_FEE_BID_INCREMENT = 0.05 // # the increment is in percentage
-// # see https://github.com/aeternity/aeternity/blob/72e440b8731422e335f879a31ecbbee7ac23a1cf/apps/aecore/src/aec_governance.erl#L272
-export const NAME_BID_TIMEOUT_BLOCKS = 480 // # ~1 day
-// # this is the max length for a domain that requires a base fee to be paid
-export const NAME_MAX_LENGTH_FEE = 31
-export const NAME_BID_MAX_LENGTH = 12 // # this is the max length for a domain to be part of a bid
-export const POINTER_KEY_BY_PREFIX = {
-  ak: 'account_pubkey',
-  ok: 'oracle_pubkey',
-  ct: 'contract_pubkey',
-  ch: 'channel'
-}
-// # https://github.com/aeternity/aeternity/blob/72e440b8731422e335f879a31ecbbee7ac23a1cf/apps/aecore/src/aec_governance.erl#L290
-// # https://github.com/aeternity/protocol/blob/master/AENS.md#protocol-fees-and-protection-times
-// # bid ranges:
-export const NAME_BID_RANGES = {
-  31: BigNumber(3).times(NAME_FEE_MULTIPLIER),
-  30: BigNumber(5).times(NAME_FEE_MULTIPLIER),
-  29: BigNumber(8).times(NAME_FEE_MULTIPLIER),
-  28: BigNumber(13).times(NAME_FEE_MULTIPLIER),
-  27: BigNumber(21).times(NAME_FEE_MULTIPLIER),
-  26: BigNumber(34).times(NAME_FEE_MULTIPLIER),
-  25: BigNumber(55).times(NAME_FEE_MULTIPLIER),
-  24: BigNumber(89).times(NAME_FEE_MULTIPLIER),
-  23: BigNumber(144).times(NAME_FEE_MULTIPLIER),
-  22: BigNumber(233).times(NAME_FEE_MULTIPLIER),
-  21: BigNumber(377).times(NAME_FEE_MULTIPLIER),
-  20: BigNumber(610).times(NAME_FEE_MULTIPLIER),
-  19: BigNumber(987).times(NAME_FEE_MULTIPLIER),
-  18: BigNumber(1597).times(NAME_FEE_MULTIPLIER),
-  17: BigNumber(2584).times(NAME_FEE_MULTIPLIER),
-  16: BigNumber(4181).times(NAME_FEE_MULTIPLIER),
-  15: BigNumber(6765).times(NAME_FEE_MULTIPLIER),
-  14: BigNumber(10946).times(NAME_FEE_MULTIPLIER),
-  13: BigNumber(17711).times(NAME_FEE_MULTIPLIER),
-  12: BigNumber(28657).times(NAME_FEE_MULTIPLIER),
-  11: BigNumber(46368).times(NAME_FEE_MULTIPLIER),
-  10: BigNumber(75025).times(NAME_FEE_MULTIPLIER),
-  9: BigNumber(121393).times(NAME_FEE_MULTIPLIER),
-  8: BigNumber(196418).times(NAME_FEE_MULTIPLIER),
-  7: BigNumber(317811).times(NAME_FEE_MULTIPLIER),
-  6: BigNumber(514229).times(NAME_FEE_MULTIPLIER),
-  5: BigNumber(832040).times(NAME_FEE_MULTIPLIER),
-  4: BigNumber(1346269).times(NAME_FEE_MULTIPLIER),
-  3: BigNumber(2178309).times(NAME_FEE_MULTIPLIER),
-  2: BigNumber(3524578).times(NAME_FEE_MULTIPLIER),
-  1: BigNumber(5702887).times(NAME_FEE_MULTIPLIER)
-}
-
-// # ref: https://github.com/aeternity/aeternity/blob/72e440b8731422e335f879a31ecbbee7ac23a1cf/apps/aecore/src/aec_governance.erl#L273
-// # name bid timeouts
-export const NAME_BID_TIMEOUTS = {
-  13: BigNumber(0),
-  12: BigNumber(NAME_BID_TIMEOUT_BLOCKS), // # 480 blocks
-  8: BigNumber(31).times(NAME_BID_TIMEOUT_BLOCKS), // # 14880 blocks
-  4: BigNumber(62).times(NAME_BID_TIMEOUT_BLOCKS) // # 29760 blocks
-}
-
-// # Tag constant for ids (type uint8)
-// # see https://github.com/aeternity/protocol/blob/master/serializations.md#the-id-type
-// # <<Tag:1/unsigned-integer-unit:8, Hash:32/binary-unit:8>>
-const ID_TAG_ACCOUNT = 1
-const ID_TAG_NAME = 2
-const ID_TAG_COMMITMENT = 3
-const ID_TAG_ORACLE = 4
-const ID_TAG_CONTRACT = 5
-const ID_TAG_CHANNEL = 6
 
 // # OBJECT tags
 // # see https://github.com/aeternity/protocol/blob/master/serializations.md#binary-serialization
@@ -156,9 +80,6 @@ const OBJECT_TAG_SOPHIA_BYTE_CODE = 70
 
 const TX_FIELD = (name, type, prefix) => [name, type, prefix]
 const TX_SCHEMA_FIELD = (schema, objectId) => [schema, objectId]
-
-const revertObject = (obj) => Object.entries(obj)
-  .reduce((acc, [key, v]) => (acc[v] = key) && acc, {})
 
 /**
  * @constant
@@ -400,24 +321,6 @@ export const TX_FEE_OTHER_GAS = (txType, txSize, { relativeTtl, innerTxSize }) =
       return new BigNumber(txSize).times(GAS_PER_BYTE)
   }
 }
-
-export const ID_TAG = {
-  account: ID_TAG_ACCOUNT,
-  name: ID_TAG_NAME,
-  commitment: ID_TAG_COMMITMENT,
-  oracle: ID_TAG_ORACLE,
-  contract: ID_TAG_CONTRACT,
-  channel: ID_TAG_CHANNEL
-}
-export const PREFIX_ID_TAG = {
-  ak: ID_TAG.account,
-  nm: ID_TAG.name,
-  cm: ID_TAG.commitment,
-  ok: ID_TAG.oracle,
-  ct: ID_TAG.contract,
-  ch: ID_TAG.channel
-}
-export const ID_TAG_PREFIX = revertObject(PREFIX_ID_TAG)
 
 const BASE_TX = [
   TX_FIELD('tag', FIELD_TYPES.int),
